requirements.txt
faiss-cpu
opencv-python
scikit-image
tensorflow
numpy
Pillow
streamlit
streamlit-cropper